Created the main Layout for the app
Used Edittext, TextView, RodioGroup with vertical orientation  and three RadioButtons also checked one of the Buttons as default.
Used the Switch with Text and used a Button as well. All that Views in the ConstraintLayout with padding 16dp.
All hardcoed strings are saved in the values/string.xml with an unique name and can be reused in the programm.